AIS

Documentation for AIS.

AIS.ConvexPolygon

AIS.Geometry

AIS.Planar

AIS.Planar

AIS.Ship

AIS.CrossProduct

AIS.InOutJudge

AIS.dist

AIS.execute

AIS.mean

AIS.rotate

Base.:+

Base.:-

Base.cos

Base.maximum

Base.minimum

AIS.ConvexPolygon — Type

ConvexPolygon is a convex polygon on planar and geographic coordinate system. ConvexPolygon.gvertex is vertexes of the convex polygon on the geographic coordinate system. ConvexPolygon.pvertex is vertexes of the convex polygon on the planar coordinate system.

struct ConvexPolygon
    gvertex::Vector{Geometry}
    pvertex::Vector{Planar}
end

ConvexPolygon(G::Vector{Geometry}) = ConvexPolygon(G,Planar.(G)) ConvexPolygon(P::Vector{Planar}) = ConvexPolygon(Geometry.(P),P)

AIS.Geometry — Type

Geometry is a coordinate on the geographic coordinate system.

struct Geometry
    longitude::Real
    latitude::Real
end

AIS.Planar — Type

Planar is a coordinate on the planar projection of Geometry with ContactPoint as a point of contanct. In a Planar Coordinate System, the units are meters.

struct Planar
    x::Real
    y::Real
end

AIS.Planar — Method

Planar(G::Geometry;ContactPoint::Geometry=DefaultRefPoint)

Return a Planar coordinate Planar(x,y) on the planar projection of G with ContactPoint as a point of contact. By default, TokyoBayRefPoint is a DefaultRefPoint.

AIS.Ship — Type

Ship is a ConvexPolygon that have a center point on planar and geographic coordinate system. Ship.Shape is the shape of the ship. It's like a home base. Ship.Geometry is a center point on geographic coordinate system. Ship.Planar is a center point on planar coordinate system.

struct Ship
    Shape::ConvexPolygon
    Geometry::Geometry
    Planar::Planar
end

TrueHeading specifies the direction of the bow. RefPoint specifies the shape of the ship. RefPoint is given in the form 'a-b-c-d' and 'a' is the length of the upper side from the anntena. 'b' is the lower size, 'c' is the left side and 'd' is the right side. edge_param specifies the sharpness of the bow of the ship. Ship(P::Planar,TrueHeading,RefPoint;edge_param=0.9)`

AIS.CrossProduct — Method

CrossProduct(A::Geometry,B::Geometry,C::Geometry) CrossProduct(A::Planar,B::Planar,C::Planar)

Get a cross product of B-A and C-A.

AIS.InOutJudge — Method

InOutJudge(A::ConvexPolygon,x::Geometry) InOutJudge(A::ConvexPolygon,x::Planar)

Return true if x is in A.

AIS.dist — Method

dist(A::Planar,B::Planar)

Get the Euclid distance of A and B.

dist(A::Planar) = dist(A,Planar(0.0,0.0))

dist(P::Vector{Planar})

Get the Euclid distance matrix of each element of P

AIS.execute — Method

execute(path)

Return the result of SQL file path execution.

AIS.mean — Method

mean(P::Vector{Planar})

Get a mean vector of P.

AIS.rotate — Method

rotate(A::Planar,θ::Real)

Get a Planar rotated A θ counterclockwise

Base.:+ — Method

Base.:+(A::Planar,B::Planar)

Get a composite vector of positional vectors A and B.

Base.:- — Method

Base.:-(A::Planar,B::Planar)

Get a composite vector of positional vectors A and -B.

Base.cos — Method

Base.cos(A::Planar,B::Planar)

Get a cosine of the angle between vectors A and B.

Base.maximum — Method

Base.maximum(A::Vector{Planar};dist=dist)

Return the element of A with the maximum distance from the origin Planar(0.0,0.0). By default, the distance is Euclid distance.

Base.minimum — Method

Base.minimum(A::Vector{Planar};dist=dist)

Return the element of A with the minimum distance from the origin Planar(0.0,0.0). By default, the distance is Euclid distance.