AIS

Documentation for AIS.

AIS.ConvexPolygon

AIS.Geometry

AIS.Planar

AIS.Planar

AIS.AngleAnchoringPoint

AIS.CrossProduct

AIS.CrossProduct

AIS.InOutJudge

AIS.InOutJudge

AIS.dist

AIS.estimateBasePoint

AIS.execute

AIS.extractEdge

AIS.mean

AIS.rotate

Base.:+

Base.:-

Base.cos

Base.maximum

Base.minimum

AIS.ConvexPolygon — Type

ConvexPolygon is a convex polygon on planar and geographic coordinate system. ConvexPolygon.gvertex is vertexes of the convex polygon on the geographic coordinate system. ConvexPolygon.pvertex is vertexes of the convex polygon on the planar coordinate system.

struct ConvexPolygon
    gvertex::Vector{Geometry}
    pvertex::Vector{Planar}
end

ConvexPolygon(G::Vector{Geometry}) = ConvexPolygon(G,Planar.(G))

AIS.Geometry — Type

Geometry is a coordinate on the geographic coordinate system.

struct Geometry
    longitude::Real
    latitude::Real
end

AIS.Planar — Type

Planar is a coordinate on the planar projection of Geometry with ContactPoint as a point of contanct. In a Planar Coordinate System, the units are meters.

struct Planar
    x::Real
    y::Real
end

AIS.Planar — Method

Planar(G::Geometry;ContactPoint::Geometry=DefaultRefPoint)

Return a Planar coordinate Planar(x,y) on the planar projection of G with ContactPoint as a point of contact. By default, TokyoBayRefPoint is a DefaultRefPoint.

AIS.AngleAnchoringPoint — Method

AngleFromBasePoint(P::Vector{Planar};d=100)

get vector of angle between the vertex of estimateBasePoint and P[i]. Strictry, Angle at the vartex V of a triangle consisting of three points V, P[i], and the mid point of triangle base.

AIS.CrossProduct — Method

CrossProduct(A::Geometry,B::Geometry,C::Geometry)

Get a cross product of B-A and C-A.

AIS.CrossProduct — Method

CrossProduct(A::Planar,B::Planar,C::Planar)

Get a cross product of B-A and C-A.

AIS.InOutJudge — Method

InOutJudge(A::ConvexPolygon,x::Geometry)

Return true if x is in A.

AIS.InOutJudge — Method

InOutJudge(A::ConvexPolygon,x::Planar)

Return true if x is in A.

AIS.dist — Method

dist(A::Planar,B::Planar)

Get the Euclid distance of A and B.

dist(A::Planar) = dist(A,Planar(0.0,0.0))

dist(P::Vector{Planar})

Get the Euclid distance matrix of each element of P

AIS.estimateBasePoint — Method

estimateBasePoint(P::Vector{Planar};d=100,returnMid=false)

Get a vertex of the isosceles triangle whose base is side of extractEdge(P::Vector{Planar}) and height is d.

Return the mid point of the triangle base if returnMid=true.

AIS.execute — Method

execute(path)

Return the result of SQL file path execution.

AIS.extractEdge — Method

extractEdge(P::Vector{Planar})

get edge of P. P[i] is an edge candigate if P[i] satisfies cos(P[i],P[i-1])<0

Edge is the pair that maximize Euclid distance in edge candigates.

AIS.mean — Method

mean(P::Vector{Planar})

Get a mean vector of P.

AIS.rotate — Method

rotate(A::Planar,θ::Real)

Get a Planar rotated A θ counterclockwise

Base.:+ — Method

Base.:+(A::Planar,B::Planar)

Get a composite vector of positional vectors A and B.

Base.:- — Method

Base.:-(A::Planar,B::Planar)

Get a composite vector of positional vectors A and -B.

Base.cos — Method

Base.cos(A::Planar,B::Planar)

Get a cosine of the angle between vectors A and B.

Base.maximum — Method

Base.maximum(A::Vector{Planar};dist=dist)

Return the element of A with the maximum distance from the origin Planar(0.0,0.0). By default, the distance is Euclid distance.

Base.minimum — Method

Base.minimum(A::Vector{Planar};dist=dist)

Return the element of A with the minimum distance from the origin Planar(0.0,0.0). By default, the distance is Euclid distance.